Add Doxygen documnetation for CTDataset (Task 6.7.3)
Added the Doxygen /doc folder to .gitignore.
Also changed how the image dimensions and number of CT layers get set (before it was just using the constant in both widget and CTDataset, now CTDataset has a constructor with the dimensions of the image and the number of layers). As a consequence, the signature of calculate- and renderDepthBuffer changed as well as the corresponding variables used in the implementation. Also marked getter functions as const.